refactor: Move all functions to card.js file and all tests to card-test.js file
