--- src/deck.js
@@ -1,9 +1,0 @@
-function createDeck(cards) {
-  return cards;
-}
-
-function countCards() {
-
-}
-
-module.exports = { createDeck, countCards };
