--- src/turn.js
@@ -1,8 +1,0 @@
-function evaluateGuess(userGuess, card) {
-  if (userGuess === card.correctAnswer) {
-    return 'correct!'
-  }
-  return 'incorrect!'
-};
-
-module.exports = { evaluateGuess }
