+++ test/Card-test.js
@@ -1,7 +1,7 @@
 const chai = require('chai');
 const expect = chai.expect;
 
-const { createCard } = require('../src/card');
+const { createCard, evaluateGuess, createDeck, countCards } = require('../src/card');
 
 describe('createCard', function() {
   it('should be a function', function() {
@@ -16,4 +16,36 @@ describe('createCard', function() {
     expect(card.answers).to.deep.equal(['object', 'array', 'function']);
     expect(card.correctAnswer).to.equal('object');
   });  
+});
+
+describe('evaluateGuess', () => {
+  it("should return 'correct!' if the guess is correct", () => {
+    const card = createCard(1, 'What allows you to define a set of related information using key-value pairs?', ['object', 'array', 'function'], 'object');
+    const userGuess = 'object';
+    const turn = evaluateGuess(userGuess, card);
+    expect(turn).to.equal('correct!');
+  });
+  it("should return 'incorrect!' if the guess is incorrect", () => {
+    const card = createCard(1, 'What allows you to define a set of related information using key-value pairs?', ['object', 'array', 'function'], 'object');
+    const userGuess = 'array';
+    const turn = evaluateGuess(userGuess, card);
+    expect(turn).to.equal('incorrect!');
+  })
+});
+
+describe('deck', () => {
+  it('should create a deck of cards', () => {
+    const card1 = createCard(1, 'What allows you to define a set of related information using key-value pairs?', ['object', 'array', 'function'], 'object');
+    const card2 = createCard(14, 'What organ is Khalid missing?', ['spleen', 'appendix', 'gallbladder'], 'gallbladder');
+    const card3 = createCard(12, 'What is Travis\'s middle name?', ['Lex', 'William', 'Fitzgerald'], 'Fitzgerald');
+    const deck = createDeck([card1, card2, card3]);
+    const deckNum = countCards(deck);
+    expect(deckNum).to.equal(3);
+  });
+  it('should work with different size decks', () => {
+    const card1 = createCard(1, 'What allows you to define a set of related information using key-value pairs?', ['object', 'array', 'function'], 'object');
+    const deck = createDeck([card1]);
+    const deckNum = countCards(deck);
+    expect(deckNum).to.equal(1);
+  });
 });
